Firebase Update:
New:
Can now save and retrieve workout information to Firebase! Awesome milestone.
Removed some calls to classes which I don't actually use.

Issues:
Can not update cards yet, the way I wrote it in SQLite beforehand means this is
going to be a pain in the ass, but its not impossible.

Next:
Enable more ways to log in so my friends can help me test this as I do
not have an Android Device.
Create a "new user" page with all the options (facebook, google, email).