General Bug fixes and finally got the CalendarView working.
Next I will try and implement the search functionality in the lists (and
maybe also add more data to the lists)
and also try and get the copy and paste feature working. Would be nice
to change the delete from onHold to it beinng onHold and then a little
bin icon showing on the action bar.